third_party/blink/web_tests/platform/generic/inspector-protocol/network/websocket-initiator-expected.txt

Initiator for Websockets check fixes http://crbug.com/457811
Test started
Network agent enabled

Initiator Type: script
Stack #0
  functionName: createSocket
  url: websocket-initiator.html
  lineNumber: 3
